requirements.txt
h5py~=3.12.1
matplotlib~=3.9.2
numpy==2.1.1
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==10.4.0
scipy==1.14.1
setuptools==75.1.0
scikit-image>=0.23

scikit-learn~=1.5.2
opencv-python~=4.10.0.84
shapely~=2.0.6
frechetdist~=0.6